+++ Code/backend/bug/views.py
@@ -53,7 +53,7 @@ def bugGetAllByUser(request, page):
     try:
         user = GeneralUser.objects.get(id = request.user.id)
         bugs = Bug.objects.filter(user = user).order_by('-createdDate')
-        pages = (bugs.count() + 3) / 4
+        pages = int((bugs.count() + 3) / 4)
         bugs = bugs[((page - 1) * 4) : (page * 4)]
         return jsons([dict(bug.body()) for bug in bugs], 0, pages)
     except GeneralUser.DoesNotExist:
@@ -61,7 +61,7 @@ def bugGetAllByUser(request, page):
 
 def bugGetAllByPage(request, page, count):
     bugs = Bug.objects.all().order_by('-createdDate')
-    pages = (bugs.count() + (count - 1)) / count
+    pages = int((bugs.count() + (count - 1)) / count)
     bugs = bugs[((page - 1) * count) : (page * count)]
 
     return jsons([dict(bug.body()) for bug in bugs], 0, pages)
@@ -74,13 +74,13 @@ def bugGetAllPageCount(request, count):
 
 def bugGetByStatusAndPage(request, status, page, count):
     bugs = Bug.objects.filter(status = status).order_by('-createdDate')
-    pages = (bugs.count() + (count - 1)) / count
+    pages = int((bugs.count() + (count - 1)) / count)
     bugs = bugs[((page - 1) * count) : (page * count)]
 
     return jsons([dict(bug.body()) for bug in bugs], 0, pages)
 
 def bugGetStatusPageCount(request, status, count):
     bugs = Bug.objects.filter(status = status)
-    pages = (bugs.count() + (count - 1)) / count
+    pages = int((bugs.count() + (count - 1)) / count)
 
     return jsons([], 0, pages)
